+++ src/low-level/imap/mailimap_oauth2.c
@@ -159,7 +159,7 @@ int mailimap_oauth2_authenticate_send(mailimap * session,
   ptr = memcpy(ptr + access_token_len, "\1\1\0", 3);
   
   /* Convert to base64 */
-  full_auth_string_b64 = encode_base64(full_auth_string, full_auth_string_len);
+  full_auth_string_b64 = encode_base64(full_auth_string, (int)full_auth_string_len);
   if (full_auth_string_b64 == NULL) {
     res = MAILIMAP_ERROR_MEMORY;
     goto free;
